Added more tests and started altering drag and drop
Drag and Drop should now allow the dragging of a draggable element around the page.  When dropped will stay where it was dropped, using absolute positioning.  If dropped within a dropZone though, it will be appended to that dropzone instead of using absolute positioning

The dropped element when appended to the dropzone doesn't look right, so some CSS needs to be used to make it take up the entire space allowed by the dropzone element.